requirements.txt
yt-dlp
pyinstaller
requests
packaging
pillow